+++ index.html
@@ -3,8 +3,6 @@
 
 <head>
     <title>An Empty World</title>
-    <link rel="stylesheet" type="text/css" href="css/style.css">
-
 </head>
 
 <body>
@@ -232,12 +230,5 @@
         <a onclick="resetData()">restart</a>
         <!-- <p>&copy; Idle Game</p> -->
 </body>
-
-<script type="module" src="js/main.js"></script>
-
-<!-- <script type="text/javascript">
-    window.glob = "inventory";
-</script> -->
-
 
 </html>
