+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+      "outDir": "./dist/",
+      "noImplicitAny": false,
+      "module": "es6",
+      "target": "es5",
+      "jsx": "react",
+      "allowJs": true,
+      "moduleResolution": "node",
+      "sourceMap": true
+    },
+    "include": ["js/**/*"],
+    "exclude": ["node_modules"]
+}
